Update Node.java
afstandVanBeginpunt type van float naar double veranderd

+++ Summatieve opdrachten/3. Dijkstra Shortest Path & Datastructuren/Node.java
@@ -5,11 +5,11 @@ import java.util.Comparator;
 public class Node implements Comparator<Node> {
 
     public int node;
-    public float afstandVanBeginpunt;
+    public double afstandVanBeginpunt;
 
     public Node() { }
 
-    public Node(int node, float afstandVanBeginpunt) {
+    public Node(int node, double afstandVanBeginpunt) {
         this.node = node;
         this.afstandVanBeginpunt = afstandVanBeginpunt;
     }
